requirements.txt
Flask==0.12.2
PyYAML==5.1
GitPython==2.1.8
docker==2.7.0
requests==2.18.4
Click==7.0
